requirements.txt
streamlit
flask
pymongo
st-star-rating
streamlit-card
openai
azure-identity
langchain-openai
langchain
pdfplumber
pymupdf
python-dotenv
flask-cors
chromadb
pypdf
langchain-community
pydantic
pydantic[email]
pandas
langchain-chroma>=0.1.2
google-cloud-storage
msal
Authlib
streamlit-cookies-manager
streamlit-js-eval
langgraph
st-cookie>=1.0